% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpcon-package.R
\docType{package}
\name{corpdiss-package}
\alias{corpdiss}
\alias{corpdiss-package}
\title{corpdiss: A Corpus-Based Model of Dissonance Perception}
\description{
This package implements a corpus-based model of dissonance perception. This model represents the idea that dissonance decreases through familiarity, and therefore perceptual dissonance should decrease in proportion to a chord's prevalence in a given musical style.
}
\author{
\strong{Maintainer}: Peter Harrison \email{pmc.harrison@gmail.com}

}
\keyword{internal}
