% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-dissonance.R
\name{corpus_dissonance}
\alias{corpus_dissonance}
\title{Corpus dissonance}
\usage{
corpus_dissonance(x, table = corpdiss::popular_1_pc_chord_type)
}
\arguments{
\item{x}{Sonority to analyse.
This should be an object created by the \code{hrep} package,
representing a pitch chord (\code{\link[hrep]{pi_chord}}),
representing a pitch-class chord (\code{\link[hrep]{pc_chord}}),
representing a pitch-class chord type (\code{\link[hrep]{pc_chord_type}}),
or a pitch-class set (\code{\link[hrep]{pc_set}}).
This object will be coerced to the same type as the table,
i.e. \code{type(table)}.}

\item{table}{Corpus dissonance table, as created by
\code{\link{corpus_dissonance_table}()}.
This table summarises chord prevalences within a corpus.
The default is \code{\link{popular_1_pc_chord_type}}.}
}
\value{
Dissonance estimate, as a numeric scalar.
}
\description{
Calculates a corpus-based estimate of the dissonance of a sonority.
}
