% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{type}
\alias{type}
\title{Type}
\arguments{
\item{x}{Object whose type should be extracted.}
}
\value{
Character scalar identifying the object's type.
An error is thrown if the type is undefined.
}
\description{
Extracts the type of an object.
}
